[Video Conversion] Add checks on files existing; Fix Help example syntax.

+++ PowerShell/function-ConvertTo-MP4andTransferToServer.ps1
@@ -13,11 +13,11 @@ function ConvertTo-MP4andTransferToServer {
     .PARAMETER Path
         String. The path to the MKV file to convert.
     .PARAMETER Destination
-        String. The destination path to transfer the MP4 file to.
+        String. The destination directory to transfer the MP4 file to.
     .PARAMETER Loop
         Switch. Optional switch to continuously loop.
     .EXAMPLE
-        > ConvertTo-MP4andTransferToServer -Path "C:\Users\${env:USERNAME}\Videos\MakeMKV\Home Videos S04*.mkv" -Destination "\\Server\Videoes\Home Videos\Series 4"
+        $ ConvertTo-MP4andTransferToServer -Path "C:\Users\${env:USERNAME}\Videos\MakeMKV\Home Videos S04*.mkv" -Destination "\\Server\Videoes\Home Videos\Series 4"
         19:25:11: TITLE: Home Videos S04 E11
         & C:\Program Files\HandBrake\HandBrakeCLI\HandBrakeCLI.exe -i C:\Users\User\Videos\MakeMKV\Home Videos S04 E11.mkv -o C:\Users\User\Videos\MakeMKV\Home Videos S04 E11.mp4 --preset "Fast 1080p30"
         . . .
@@ -26,7 +26,7 @@ function ConvertTo-MP4andTransferToServer {
         19:33:05: Moving converted MP4 to Done: C:\Users\User\Videos\MakeMKV\Home Videos S04 E11.mp4
         19:33:05: Start-Sleep -Seconds 900 (Next kick-off: 19:48:05)
     .EXAMPLE
-        > ConvertTo-MP4andTransferToServer -Path "C:\Users\${env:USERNAME}\Videos\MakeMKV\Wedding*.mkv" -Destination "\\Server\Home Movies"
+        $ ConvertTo-MP4andTransferToServer -Path "C:\Users\${env:USERNAME}\Videos\MakeMKV\Wedding*.mkv" -Destination "\\Server\Home Movies"
         19:25:11: TITLE: Wedding Day
         & C:\Program Files\HandBrake\HandBrakeCLI\HandBrakeCLI.exe -i C:\Users\User\Videos\MakeMKV\Wedding Day.mkv -o C:\Users\User\Videos\MakeMKV\Wedding Day.mp4 --preset "Fast 1080p30"
         . . .
@@ -53,7 +53,19 @@ function ConvertTo-MP4andTransferToServer {
     $NonEncodeSleep  = 30  # 30 seconds
     $HandBrakeCliExe = "C:\Program Files\HandBrake\HandBrakeCLI\HandBrakeCLI.exe"
 
-    # TODO: Check for files and folders.
+    # Check that necessary files and folders exist.
+    if ($null -eq $Path -or (-not (Test-Path -Path $Path))) {
+        Write-Error "Path is null or empty."
+        return
+    }
+    if ($null -eq $Destination -or (-not (Test-Path -Path $Destination))) {
+        Write-Error "Destination is null or empty."
+        return
+    }
+    if (-not (Test-Path -Path $HandBrakeCliExe)) {
+        Write-Error "HandBrake CLI is null or empty."
+        return
+    }
 
     do {
         $Item = Get-Item -Path $Path | Sort-Object -Property LastWriteTime | Select-Object -First 1
@@ -66,10 +78,13 @@ function ConvertTo-MP4andTransferToServer {
             Write-Host "$(Get-Date -Format "HH:mm:ss"): TITLE: $ItemTitle" -ForegroundColor Green
             Write-Host "& $HandBrakeCLIExe -i $ItemFullName -o ${ItemDirectoryName}\${ItemTitle}.mp4 --preset `"Fast 1080p30`"" -ForegroundColor Green
             & $HandBrakeCLIExe -i $ItemFullName -o ${ItemDirectoryName}\${ItemTitle}.mp4 --preset "Fast 1080p30"
+            if (-not (Test-Path -Path ${ItemDirectoryName}\${ItemTitle}.mp4)) {
+                Write-Error "Failed to convert MKV to MP4. Exiting."
+                return
+            }
             Write-Host ""
-            # TODO: Check convert success.
 
-            Write-Host "$(Get-Date -Format "HH:mm:ss"): Moving original MKV to Done: $ItemFullName" -ForegroundColor Green
+            Write-Host "$(Get-Date -Format "HH:mm:ss"): Moving original MKV to Done: $ItemFullName" -ForegroundColor Green # TODO: Delete original MKV file instead.
             Move-Item -Path $ItemFullName -Destination $ItemDoneDirectory
             Write-Host ""
 
@@ -77,8 +92,7 @@ function ConvertTo-MP4andTransferToServer {
             Start-BitsTransfer -Source ${ItemDirectoryName}\${ItemTitle}.mp4 -Destination $Destination -DisplayName 'Transfer video file' -Description 'Transfer Video File..'
 
             Write-Host "$(Get-Date -Format "HH:mm:ss"): Moving converted MP4 to Done: ${ItemDirectoryName}\${ItemTitle}.mp4" -ForegroundColor Green
-            Move-Item -Path ${ItemDirectoryName}\${ItemTitle}.mp4 -Destination $ItemDoneDirectory
-            # TODO: Delete original MKV file / Converted MP4 file if successful
+            Move-Item -Path ${ItemDirectoryName}\${ItemTitle}.mp4 -Destination $ItemDoneDirectory # TODO: Delete MP4 file instead.
             Write-Host ""
 
             if ($Loop) {
